+++ src/types/environment.d.ts
@@ -1,0 +1,9 @@
+export type {}
+
+declare global {
+  namespace NodeJS {
+    interface ProcessEnv {
+      NODE_ENV: 'development' | 'production'
+    }
+  }
+}