+++ CHANGELOG.md
@@ -7,11 +7,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- `public` folder is served as a static webserver root folder
+- user can configure mime-type per extension
+- user can configure headers per resource
+
+### Versioning
+
+- Development dependencies updated
+
 ## [2.0.4] - 2023-02-17
 
 ### Fixed
 
 - cli fix: from `busDiscriminator` to `eventBus` pool even when other properties are declared
+- headers: use `/public/index.html` headers when 200 and url is `/public` or `/public/`
 
 ### Versioning